+++ extensions/spark/kyuubi-spark-authz/src/main/scala/org/apache/kyuubi/plugin/spark/authz/ranger/RuleAuthorization.scala
@@ -17,22 +17,19 @@
 
 package org.apache.kyuubi.plugin.spark.authz.ranger
 
-import java.util
-
 import scala.collection.JavaConverters._
-import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
 
-import org.apache.commons.collections.CollectionUtils
 import org.apache.ranger.plugin.policyengine.RangerAccessRequest
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.rules.Rule
 import org.apache.spark.sql.catalyst.trees.TreeNodeTag
 
-import org.apache.kyuubi.plugin.spark.authz.{AccessControlException, ObjectType, _}
+import org.apache.kyuubi.plugin.spark.authz._
 import org.apache.kyuubi.plugin.spark.authz.ObjectType._
 import org.apache.kyuubi.plugin.spark.authz.ranger.RuleAuthorization.KYUUBI_AUTHZ_TAG
+import org.apache.kyuubi.plugin.spark.authz.ranger.SparkRangerAdminPlugin._
 import org.apache.kyuubi.plugin.spark.authz.util.AuthZUtils._
 
 class RuleAuthorization(spark: SparkSession) extends Rule[LogicalPlan] {
@@ -48,9 +45,6 @@ class RuleAuthorization(spark: SparkSession) extends Rule[LogicalPlan] {
 object RuleAuthorization {
 
   val KYUUBI_AUTHZ_TAG = TreeNodeTag[Boolean]("__KYUUBI_AUTHZ_TAG")
-
-  val CONF_FULL_ACCESS_CHECK_ENABLE =
-    s"ranger.plugin.${SparkRangerAdminPlugin.getServiceType}.enable.full.access.check"
 
   def checkPrivileges(spark: SparkSession, plan: LogicalPlan): Unit = {
     val auditHandler = new SparkRangerAuditHandler
@@ -90,68 +84,20 @@ object RuleAuthorization {
         case _ => accessReqs += Seq(request)
       }
     }
-    val isEnabledFullAccessCheck =
-      SparkRangerAdminPlugin.getConfig.getBoolean(CONF_FULL_ACCESS_CHECK_ENABLE, false)
-    if (isEnabledFullAccessCheck) {
-      verify(
+
+    if (isEnabledFullAccessCheck()) {
+      SparkRangerAdminPlugin.verify(
         accessReqs.toStream.flatMap(_.toStream).asJava,
         auditHandler,
-        isEnabledFullAccessCheck = true)
+        enabledFullAccessViolationMsg = true)
     } else {
       accessReqs.foreach {
         reqList =>
-          verify(reqList.asJava, auditHandler, isEnabledFullAccessCheck = false)
+          SparkRangerAdminPlugin.verify(
+            reqList.asJava,
+            auditHandler,
+            enabledFullAccessViolationMsg = false)
       }
-    }
-  }
-
-  @throws[AccessControlException]
-  private def verify(
-      requests: util.List[RangerAccessRequest],
-      auditHandler: SparkRangerAuditHandler,
-      isEnabledFullAccessCheck: Boolean): Unit = {
-    if (CollectionUtils.isEmpty(requests)) {
-      return
-    }
-
-    val results = SparkRangerAdminPlugin.isAccessAllowed(requests, auditHandler)
-
-    val disallowedReqs = requests.asScala.zip(results.asScala).filter {
-      case (_, result) => result != null && !result.getIsAllowed
-    } map { case (req, _) => req }
-    if (disallowedReqs.isEmpty) {
-      return
-    }
-
-    if (!isEnabledFullAccessCheck) {
-      val req = disallowedReqs.headOption.orNull
-      if (req != null) {
-        if (!isEnabledFullAccessCheck) {
-          throw new AccessControlException(
-            s"Permission denied: user [${req.getUser}]" +
-              s" does not have [${req.getAccessType}]" +
-              s" privilege on [${req.getResource.getAsString}]")
-        }
-      }
-    } else {
-      val accessType2ResMap: mutable.Map[String, mutable.Set[String]] =
-        mutable.SortedMap()
-      disallowedReqs.foreach { req =>
-        {
-          val resourceSet = accessType2ResMap.getOrElseUpdate(
-            req.getAccessType,
-            mutable.SortedSet[String]())
-          resourceSet += req.getResource.getAsString
-        }
-      }
-
-      val user: String = disallowedReqs.head.getUser
-      val privilegeErrorMsg = accessType2ResMap.map {
-        case (accessType, resSet) =>
-          s"[${accessType}] privilege on [${resSet.mkString(",")}]"
-      }.mkString(", ")
-      throw new AccessControlException(
-        s"Permission denied: user [${user}] does not have ${privilegeErrorMsg}")
     }
   }
 }
